Allows advanced users to create deployments and control those deployed desktops

+++ webapp/webapp/webapp/lib/admin_api.py
@@ -84,6 +84,13 @@ class isardAdmin():
                 res_stopped=r.table(table).get_all(r.args(domains_stopped)).update({'status':'Starting'}).run(db.conn)
                 res_started=r.table(table).get_all(r.args(domains_started)).update({'status':'Stopping'}).run(db.conn)
                 return True
+            if action == 'toggle_visible':
+                domains_shown=self.multiple_check_field(table,'tag_visible',True,ids)
+                domains_hidden=self.multiple_check_field(table,'tag_visible',False,ids)
+                res_shown=r.table(table).get_all(r.args(domains_hidden)).update({'tag_visible':True}).run(db.conn)
+                res_hidden=r.table(table).get_all(r.args(domains_shown)).filter({'status':'Started'}).update({'status':'Stopping'}).run(db.conn)
+                res_hidden=r.table(table).get_all(r.args(domains_shown)).update({'tag_visible':False,'viewer':False}).run(db.conn)
+                return True
             if action == 'delete':
                 domains_deleting=self.multiple_check_field(table,'status','Deleting',ids)
                 res=r.table(table).get_all(r.args(domains_deleting)).delete().run(db.conn) 
@@ -405,7 +412,7 @@ class isardAdmin():
             for u in users:
                 u.update({"kind":"user","user":u['id']})
 
-            group_desktops = list(r.table("domains").get_all(group_id, index='group').filter({'kind': 'desktop'}).pluck('id','name','kind','user','status','parents').run(db.conn))
+            deployment_desktops = list(r.table("domains").get_all(group_id, index='group').filter({'kind': 'desktop'}).pluck('id','name','kind','user','status','parents').run(db.conn))
             group_templates = list(r.table("domains").get_all(r.args(['base','public_template','user_template']),index='kind').filter({'group':group_id}).pluck('id','name','kind','user','status','parents').run(db.conn))
             derivated = []
             for gt in group_templates:
@@ -413,7 +420,7 @@ class isardAdmin():
                 derivated = derivated + list(r.table('domains').pluck('id','name','kind','user','status','parents').filter(lambda derivates: derivates['parents'].contains(id)).run(db.conn))
                 #templates = [t for t in derivated if t['kind'] != "desktop"]
                 #desktops = [d for d in derivated if d['kind'] == "desktop"]
-        domains = groups+users+group_desktops+group_templates+derivated
+        domains = groups+users+deployment_desktops+group_templates+derivated
         return [i for n, i in enumerate(domains) if i not in domains[n + 1:]] 
     
     def group_delete(self,group_id):
